Create app.js
This is a JavaScript file that interacts with a deployed smart contract on the Ethereum blockchain. The smart contract is a charity contract that accepts donations in ERC20 tokens.

The code utilizes the Web3 library, which is a collection of libraries that allow JavaScript to interact with Ethereum nodes using HTTP or WebSocket connections.

The first few lines of the code set the ABI (Application Binary Interface) for the charity smart contract, the address of the deployed contract, and the web3 provider URL that will be used to communicate with the Ethereum network.

The code defines several functions that allow users to interact with the smart contract. The `donate()` function allows users to donate ERC20 tokens to the contract, the `getBalance()` function retrieves the balance of the contract, the `withdrawFunds()` function allows the contract owner to withdraw funds from the contract, the `addToken()` function adds a new ERC20 token to the contract, the `getDonationHistory()` function retrieves the donation history for a specific user, the `addToWhitelist()` function adds an address to the whitelist, and the `removeFromWhitelist()` function removes an address from the whitelist.

The functions utilize the Web3 library to call the appropriate functions on the smart contract, passing in any necessary parameters such as the amount to donate or the address to add to the whitelist. The functions also require specific addresses to execute them such as the user's address, contract owner's address, or the address to be added or removed from the whitelist.

Before calling the `donate()` function, it first checks whether the ERC20 token being donated is an approved token for the contract. If the token is not approved, an error is thrown.

This code can be used as a starting point for building a front-end application that interacts with the charity smart contract. However, it is important to properly audit the code and test it thoroughly before using it in a production environment to ensure that it is secure and functions as intended.